Adjust ActionsPreprocessing with actions serialization
